Text.Pandoc.Definition

Description

Definition of Pandoc data structure for format-neutral representation of documents.

Documentation

data Pandoc

Constructors

Pandoc Meta [Block]

newtype Meta

Metadata for the document: title, authors, date.

Constructors

Meta

Fields

unMeta :: Map String MetaValue

data MetaValue

Constructors

MetaMap (Map String MetaValue)

MetaList [MetaValue]

MetaBool Bool

MetaString String

MetaInlines [Inline]

MetaBlocks [Block]

nullMeta :: Meta

isNullMeta :: Meta -> Bool

lookupMeta :: String -> Meta -> Maybe MetaValue

Retrieve the metadata value for a given key.

docTitle :: Meta -> [Inline]

Extract document title from metadata; works just like the old docTitle.

docAuthors :: Meta -> [[Inline]]

Extract document authors from metadata; works just like the old docAuthors.

docDate :: Meta -> [Inline]

Extract date from metadata; works just like the old docDate.

data Block

Block element.

Constructors

Plain [Inline]
Plain text, not a paragraph

Para [Inline]
Paragraph

CodeBlock Attr String
Code block (literal) with attributes

RawBlock Format String
Raw block

BlockQuote [Block]
Block quote (list of blocks)

OrderedList ListAttributes [[Block]]
Ordered list (attributes and a list of items, each a list of blocks)

BulletList [[Block]]
Bullet list (list of items, each a list of blocks)

DefinitionList [([Inline], [[Block]])]
Definition list Each list item is a pair consisting of a term (a list of inlines) and one or more definitions (each a list of blocks)

Header Int Attr [Inline]
Header - level (integer) and text (inlines)

HorizontalRule
Horizontal rule

Table [Inline] [Alignment] [Double] [TableCell] [[TableCell]]
Table, with caption, column alignments (required), relative column widths (0 = default), column headers (each a list of blocks), and rows (each a list of lists of blocks)

Div Attr [Block]
Generic block container with attributes

Null
Nothing

data Inline

Inline elements.

Constructors

Str String
Text (string)

Emph [Inline]
Emphasized text (list of inlines)

Strong [Inline]
Strongly emphasized text (list of inlines)

Strikeout [Inline]
Strikeout text (list of inlines)

Superscript [Inline]
Superscripted text (list of inlines)

Subscript [Inline]
Subscripted text (list of inlines)

SmallCaps [Inline]
Small caps text (list of inlines)

Quoted QuoteType [Inline]
Quoted text (list of inlines)

Cite [Citation] [Inline]
Citation (list of inlines)

Code Attr String
Inline code (literal)

Space
Inter-word space

LineBreak
Hard line break

Math MathType String
TeX math (literal)

RawInline Format String
Raw inline

Link [Inline] Target
Hyperlink: text (list of inlines), target

Image [Inline] Target
Image: alt text (list of inlines), target

Note [Block]
Footnote or endnote

Span Attr [Inline]
Generic inline container with attributes

data Alignment

Alignment of a table column.

Constructors

AlignLeft

AlignRight

AlignCenter

AlignDefault

type ListAttributes = (Int, ListNumberStyle, ListNumberDelim)

List attributes.

data ListNumberStyle

Style of list numbers.

Constructors

DefaultStyle

Example

Decimal

LowerRoman

UpperRoman

LowerAlpha

UpperAlpha

data ListNumberDelim

Delimiter of list numbers.

Constructors

DefaultDelim

Period

OneParen

TwoParens

newtype Format

Formats for raw blocks

Constructors

Format String

type Attr = (String, [String], [(String, String)])

Attributes: identifier, classes, key-value pairs

nullAttr :: Attr

type TableCell = [Block]

Table cells are list of Blocks

data QuoteType

Type of quotation marks to use in Quoted inline.

Constructors

SingleQuote

DoubleQuote

type Target = (String, String)

Link target (URL, title).

data MathType

Type of math element (display or inline).

Constructors

DisplayMath

InlineMath

data Citation

Constructors

Citation

Fields

citationId :: String

citationPrefix :: [Inline]

citationSuffix :: [Inline]

citationMode :: CitationMode

citationNoteNum :: Int

citationHash :: Int

data CitationMode

Constructors

AuthorInText

SuppressAuthor

NormalCitation